+++ charts/the-web-cloud-platform/pipeline/publish.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env sh
 set -e
 
-helm push . http://127.0.0.1:58080
+helm chart save . "localhost:55000/webdock/the-web-cloud-platform:0.0.0"
+helm chart push "localhost:55000/webdock/the-web-cloud-platform:0.0.0"