+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Store Helm charts within OCI registry.
 - Use `node-fetch` instead of `Axios`.
 
 ### Fixed